Modularize 2021 plugin
