+++ plugin-2021/src/game/constants.rs
@@ -1,0 +1,1 @@
+pub const BOARD_SIZE: usize = 20;
