+++ plugin-2021/src/game/mod.rs
@@ -1,0 +1,25 @@
+mod board;
+mod color;
+mod constants;
+mod coordinates;
+mod field;
+mod game_state;
+mod r#move;
+mod piece_shape;
+mod piece;
+mod player;
+mod rotation;
+mod team;
+
+pub use board::*;
+pub use color::*;
+pub use constants::*;
+pub use coordinates::*;
+pub use field::*;
+pub use game_state::*;
+pub use r#move::*;
+pub use piece_shape::*;
+pub use piece::*;
+pub use player::*;
+pub use rotation::*;
+pub use team::*;
